refactor(test): mirror RTDB worker client families
